fix(track): widen rating column to BIGINT for sqlx i64 decode
CI caught a sqlx 0.9 decode error on the PATCH round-trip test:
SMALLINT cannot be narrowed to `Option<i64>` (the type on
`waveflow_core::domain::track::TrackRow.rating`), so the `RETURNING
… rating` projection from `UPDATE track …` returned a runtime
type-mismatch and the handler 500'd.

Earlier tests passed because the row's `rating` was still NULL
(create + list + get with no PATCH) — Option<None> doesn't trigger
the narrowing, so the bug only surfaced on the first PATCH that
actually wrote a rating.

Widening to BIGINT matches every other numeric column on this
table (file_size, duration_ms, …) and lets the i64 read+write path
stay cast-free at both ends. The CHECK (rating BETWEEN 0 AND 255)
constraint is preserved — defense in depth on top of the
TrackUpdate.rating: Option<u8> type-level guarantee from
waveflow-core. The 6-byte storage delta vs SMALLINT isn't worth a
::bigint cast on every read site.

The migration hasn't merged yet (still on this PR's branch), so an
in-place edit doesn't violate the "migrations are immutable once
merged" rule.

+++ migrations/20260530000003_track.sql
@@ -49,12 +49,17 @@ CREATE TABLE track (
     -- this repo.
     added_at        BIGINT      NOT NULL,
 
-    -- Raw POPM byte (0-255). SMALLINT is exactly enough for the
-    -- 8-bit range; the CHECK is defense in depth on top of the
-    -- `TrackUpdate.rating: Option<u8>` type-level guarantee in
-    -- waveflow-core. Out-of-range payloads can't be persisted even
-    -- if a future handler bypassed the type.
-    rating          SMALLINT    CHECK (rating BETWEEN 0 AND 255),
+    -- Raw POPM byte (0-255). BIGINT (not SMALLINT) because the
+    -- waveflow-core `TrackRow.rating: Option<i64>` projection
+    -- decodes the column directly — sqlx 0.9 refuses the SMALLINT
+    -- → i64 narrowing during RETURNING, and an explicit `::bigint`
+    -- cast on every read site would be more friction than the
+    -- 6-byte storage saving is worth. The CHECK constraint is
+    -- defense in depth on top of the `TrackUpdate.rating:
+    -- Option<u8>` type-level guarantee from waveflow-core — even
+    -- a future handler that bypasses the type can't persist a
+    -- value outside the POPM range.
+    rating          BIGINT      CHECK (rating BETWEEN 0 AND 255),
 
     UNIQUE (library_id, file_path)
 );
